INS-473 add double-check lock for creating new connections
